+++ lib/data/api_calls/ethereum_api.dart
@@ -1,9 +1,10 @@
 import 'dart:math';
-import 'package:flutter_app/api_calls/base_wallet_api.dart';
+
+import 'package:flutter_app/data/api_calls/base_wallet_api.dart';
 import 'package:flutter_app/global.dart';
-import 'package:flutter_app/models/balances.dart';
-import 'package:flutter_app/models/ethereum_wallet.dart';
-import 'package:flutter_app/models/emeris_wallet.dart';
+import 'package:flutter_app/data/model/balances.dart';
+import 'package:flutter_app/data/model/emeris_wallet.dart';
+import 'package:flutter_app/data/model/ethereum_wallet.dart';
 import 'package:http/http.dart';
 import 'package:wallet_core/wallet_core.dart';
