+++ lib/data/model/ethereum_wallet.dart
@@ -1,5 +1,5 @@
-import 'package:flutter_app/models/emeris_wallet.dart';
-import 'package:flutter_app/models/wallet_type.dart';
+import 'package:flutter_app/data/model/emeris_wallet.dart';
+import 'package:flutter_app/data/model/wallet_type.dart';
 import 'package:web3dart/web3dart.dart';
 
 class EthereumWallet implements EmerisWallet {
