+++ lib/data/model/wallet_type.dart
