+++ lib/dependency_injection/app_component.dart
@@ -1,0 +1,6 @@
+import 'package:get_it/get_it.dart';
+
+final getIt = GetIt.instance;
+
+/// registers all the dependencies in dependency graph in get_it package
+void configureDependencies() {}
