+++ lib/views/wallet_list.dart
@@ -1,7 +1,7 @@
 import 'package:clipboard/clipboard.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_app/models/emeris_wallet.dart';
-import 'package:flutter_app/models/wallet_type.dart';
+import 'package:flutter_app/data/model/emeris_wallet.dart';
+import 'package:flutter_app/data/model/wallet_type.dart';
 import 'package:flutter_app/views/wallet_details_page.dart';
 
 import '../global.dart';
