+++ pubspec.yaml
@@ -35,6 +35,7 @@ dependencies:
   encrypt: ^5.0.0
   web3dart: ^2.1.3
   bip39: ^1.0.6
+  get_it: ^7.1.3
   wallet_core:
     git:
       url: https://github.com/wal33d006/wallet_core.git
