+++ test/transaction_hosted_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter_app/api_calls/cosmos_api.dart';
-import 'package:flutter_app/api_calls/faucet_api.dart';
+import 'package:flutter_app/data/api_calls/cosmos_api.dart';
+import 'package:flutter_app/data/api_calls/faucet_api.dart';
 import 'package:flutter_app/global.dart';
 import 'package:flutter_test/flutter_test.dart';
